requirements.txt
Pillow==11.1.0
phonenumbers==9.0.3
django-phonenumber-field==8.0.0
babel==2.17.0
setuptools==78.1.0
requests~=2.32.3
python-dotenv==1.1.0
colorama~=0.4.6
django-q2==1.7.6
icalendar~=6.1.3
